test_vertex_attribute_double_types
